requirements.txt
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
debtcollector==3.0.0
Deprecated==1.2.14
gitdb==4.0.10
GitPython==3.1.37
idna==3.6
iso8601==2.1.0
keystoneauth1==5.6.0
msgpack==1.0.8
netaddr==1.2.1
netifaces==0.11.0
os-service-types==1.7.0
oslo.config==9.4.0
oslo.i18n==6.3.0
oslo.serialization==5.4.0
oslo.utils==7.1.0
packaging==23.2
pbr==6.0.0
pycparser==2.21
PyGithub==2.2.0
PyJWT==2.8.0
PyNaCl==1.5.0
pyparsing==3.1.1
python-keystoneclient==5.4.0
python-swiftclient==4.5.0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
rfc3986==2.0.0
smmap==5.0.1
stevedore==5.2.0
typing-extensions==4.0.1
urllib3==2.2.1
wrapt==1.16.0
